feat(node)!: rewrite the rollups node on one engine
The rollups node becomes a single crate: worker modules synchronize
through the storage module (the only SQL surface), and one
spec-oracled engine (src/engine: structure, stf, ruler, and the
dispute facade) serves both the runner's forward schedule and the
dispute path's random access.

- One leaf semantics: process_input and the second geometry
  implementation are gone. The runner schedules the engine's
  collect() per input window; the window-root quartet row is the
  only durable level-0 artifact; settlement and dispute serving
  read the same rows (frontier fold, strict and prefix-bounded).
- The boundary store (stage+rename, write-once cells, clone-based
  revert) carries both regimes' revert insurance; corruption
  tripwires panic through the loud process exit, never retry.
- Revert on RX_REJECTED only, matching AdvanceStatus.sol - the
  chain never reverts an exception.
- The staged settlement protocol: the epoch manager drives sentry
  claim, stage, and accept per tick, each step guarded and
  idempotent; a sentry always claims the locally computed
  post-epoch hash, never the staged value. settlement_info carries
  final_state, captured from the boundary store at roll (the new
  epoch's initial boundary IS the post-epoch state).
- Absorbs prt/client-rs (the tournament reader/sender/gc and the
  hero's react loop, traced against the contracts - the audit
  ledger records verdicts and clean checks) and common-rs (merkle,
  arithmetic, and kms become node modules; the node was their only
  consumer).
- Chain recordings - raw devnet log ranges captured after e2e
  disputes settle - are committed fixtures for the tournament
  fold, decoded through the production bindings.

Replaces the epoch-manager/state-manager/blockchain-reader/
prt-node crates, prt/client-rs, and common-rs.

+++ machine/rust-bindings/cartesi-machine/src/constants.rs
@@ -11,10 +11,10 @@
 pub mod machine {
     use cartesi_machine_sys::*;
     // pub const CYCLE_MAX: u64 = CM_MCYCLE_MAX as u64;
-    pub const HASH_SIZE: u32 = CM_HASH_SIZE as u32;
-    pub const HASH_TREE_LOG2_WORD_SIZE: u32 = CM_HASH_TREE_LOG2_WORD_SIZE as u32;
-    pub const HASH_TREE_LOG2_PAGE_SIZE: u32 = CM_HASH_TREE_LOG2_PAGE_SIZE as u32;
-    pub const HASH_TREE_LOG2_ROOT_SIZE: u32 = CM_HASH_TREE_LOG2_ROOT_SIZE as u32;
+    pub const HASH_SIZE: u32 = CM_HASH_SIZE;
+    pub const HASH_TREE_LOG2_WORD_SIZE: u32 = CM_HASH_TREE_LOG2_WORD_SIZE;
+    pub const HASH_TREE_LOG2_PAGE_SIZE: u32 = CM_HASH_TREE_LOG2_PAGE_SIZE;
+    pub const HASH_TREE_LOG2_ROOT_SIZE: u32 = CM_HASH_TREE_LOG2_ROOT_SIZE;
 }
 
 pub mod ar {
